fix bug for basic_calculator and add ut case

+++ c++/Solutions/basic_calculator.hpp
@@ -9,7 +9,7 @@ public:
     int basic_calculator(std::string pattern);
    
 private:
-    int calculator_no_parenthes(std::string pattern);
+    long long calculator_no_parenthes(std::string pattern);
     std::string strReplace(std::string sentence,
         std::string src,
         std::string dst);
